Move validateFileSafety into the security folder
